architecture(dream-cycle 2026-09-07): extract pattern-store from agentdb-adapter God-module
agentdb-adapter.ts (1483 lines) still mixes 8+ unrelated bounded contexts.
Continuing the extraction the 2026-09-02 architecture night started (and
whose own report named this exact next candidate), pattern-store
(RuclipPatternNamespace/storePattern/PatternSearchResult/searchPatterns,
~36 lines, single dependency on bridge-client.ts's callTool) moves to its
own module store/pattern-store.ts, re-exported from agentdb-adapter.ts for
zero call-site changes.

Deliberately the smallest remaining self-contained extraction, not the
largest available (heartbeat-schedule, ~180 lines): 0 of the 4 dream-cycle
PRs opened to date show merged:true via the GitHub API, so tonight biases
toward the most trivially reviewable diff per the routine's own
learning-signal rule.

Baseline (5c3ea0d, main tip): 321/321 tests, tsc --strict clean.
Candidate: 321/321 tests, tsc --strict clean. agentdb-adapter.ts
1483->1459 lines; new pattern-store.ts +50 lines; diff touches exactly two
files. Full evidence, hypothesis, evaluation receipt, reward-hack check,
and witness stamp in docs/dream-cycle/2026-09-07-architecture-report.md.

Human review required before merge — this is not self-promotion.
