+++ docs/dream-cycle/LEDGER.md
@@ -9,3 +9,4 @@ One row per run, appended only, never rewritten.
 | 2026-09-03 | performance | `checkOperatingBudget` (Gate 2, runs on every heartbeat fire) fetched N tracked session-cost keys via N sequential `memory_retrieve` round trips; replaced with a bounded worker-pool fan-out (`mapWithConcurrency`, cap 25) — output unchanged, empirically proven sequential-vs-concurrent both ways (parent fails the new concurrency regression test, candidate passes) | [#10](https://github.com/ruvnet/ruClip/issues/10) | [#11](https://github.com/ruvnet/ruClip/pull/11) | yes | ACCEPT | 155ms→32ms measured (N=6, mocked 25ms/call, 4.8x); synthetic N=200 sweep 1600ms→67ms at cap 25 (23.9x); 320→321 tests, 0 regressions | `144a11f1fd079317b6caed4a2974d5305b8fd766989ea2567f37fb493263f81d` | #8/PR#9: OPEN, still unmerged (0 of last 2 dream-cycle PRs merged as of tonight — bias toward small, reviewable candidates applied) |
 | 2026-09-05 | correctness | `wouldCreateCycle`/`graphNeighbors` (`agentdb-adapter.ts`) parsed the real `agentdb_graph-query` MCP tool's k-hop response as `{nodes:[{id}]}`; the real, exact-pinned tool (`@claude-flow/cli@3.38.20`, confirmed in this repo's own `node_modules`) returns `{results:[{nodeId}]}` — cycle prevention and `getChildIssueIds`/`getBlockerIssueIds` have been silent no-ops against the real bridge since written. Fixed the field parsing; corrected every test mock (5 call sites, 4 files) from the same wrong fictional shape to match. A first hypothesis (widen the hardcoded `depth: 5` k-hop probe) was implemented, passed 323/323, then REJECTED by an independent critic who read the real tool source and found `depth` is server-clamped to `complexityBudget.maxDepth` (default 5) regardless — a clean rejection that led directly to tonight's actual finding | [#12](https://github.com/ruvnet/ruClip/issues/12) | [#13](https://github.com/ruvnet/ruClip/pull/13) | yes | ACCEPT | Baseline (mocks corrected, code unfixed) 316/320 (4 genuine behavioral failures); candidate 320/320, 0 regressions, `harness:bench-verify` hash unchanged | `00af13527b925842621f8252b5faea310ae7750fd4062f990ac66915195a03a3` | #8/PR#9 and #10/PR#11: both still OPEN/unmerged (2026-09-04 developer-experience night appears to have not run at all — no branch/issue/PR found; not investigable further from inside this session, flagged for the routine owner) |
 | 2026-09-03 | performance | `checkOperatingBudget` (Gate 2, runs on every heartbeat fire) fetched N tracked session-cost keys via N sequential `memory_retrieve` round trips; replaced with a bounded worker-pool fan-out (`mapWithConcurrency`, cap 25) — output unchanged, empirically proven sequential-vs-concurrent both ways (parent fails the new concurrency regression test, candidate passes) | LOCAL (no `gh`/gist tool this session — see report) | LOCAL (see `docs/dream-cycle/2026-09-03-performance-report.md`; branch `dream/2026-09-03-performance` not yet pushed as of this row) | yes | ACCEPT | 155ms→32ms measured (N=6, mocked 25ms/call, 4.8x); synthetic N=200 sweep 1600ms→67ms at cap 25 (23.9x); 320→321 tests, 0 regressions | `144a11f1fd079317b6caed4a2974d5305b8fd766989ea2567f37fb493263f81d` | #8/PR#9: OPEN, still unmerged (0 of last 2 dream-cycle PRs merged as of tonight — bias toward small, reviewable candidates applied) |
+| 2026-09-07 | architecture | `agentdb-adapter.ts` (1483 lines, still the God-module) — extracted the pattern-store bounded context (`RuclipPatternNamespace`/`storePattern`/`PatternSearchResult`/`searchPatterns`, ~36 lines, single dependency on `bridge-client.ts`'s `callTool`) into `store/pattern-store.ts`, re-exported for zero call-site changes. This is the exact candidate the 09-02 architecture report's own "Next steps" named; chosen over the larger `heartbeat-schedule` candidate (~180 lines) specifically because 0/4 dream-cycle PRs opened to date show `merged:true` via the GitHub API — biased to the smallest reviewable slice per the routine's own learning-signal rule | [#8 still open](https://github.com/ruvnet/ruClip/issues/8) (09-02, unresolved); new finding filed as issue (see PR) | draft PR, branch `dream/2026-09-07-architecture` | yes | ACCEPT | `agentdb-adapter.ts` 1483→1459 lines (-24); new `pattern-store.ts` +50 lines; 321/321 tests both sides, 0 regressions; diff touches exactly 2 files | `42457282a4acff8840c09fb049dcde96909e2ddd939751cf49adea2c807912c1` | PR#9 (09-02): OPEN 5 nights, unmerged. PR#11 (09-03) and PR#13 (09-05): `state:closed, merged:false` via API yet their commits (`5c3ea0d`, `7e889d0`) are on `main` — landed by a mechanism outside this session's visibility, not a GitHub-recorded merge. PR#15 (09-06, security): OPEN, unreviewed. 0/4 dream-cycle PRs opened show `merged:true` — human-review bottleneck flagged to routine owner in tonight's report, not a candidate-quality problem |
